Fix initialization of InteractionModeManager.
The InteractionModeManager was only initialized when addActionsToToolbar
was called for scene. However it is not always guaranteed (e.g. in Apps
that are not BALLView) that this method is actually called. To fix this
we directly initialize InteractionModeManager.

For the initialization to succeed we need to guarantee, that an instance
of MainControl is registered as there are lots of calls to
getMainControl(). This requires us to call registerThis on the MainControl
instance as soon as possible.